refactor: ♻️ split large service files into directory modules and extract user_repository (#330, #325)
