+++ backend/src/repositories/mod.rs
@@ -1,1 +1,2 @@
 pub mod preview_repository;
+pub mod user_repository;
